+++ capps/include/Mts/sequential_mts.hpp
@@ -4,11 +4,11 @@
 
 
 /* This function computes the maximum prefix sum of an array with superaccumulators */
-void sequential_mts_superacc(double*,int,double*,int*);
+void sequential_mts_superacc(double*,long,double*,long*);
 
 /* This function computes the maximum prefix sum of an array with doubles */
-void sequential_mts_double(double*,int,double*,int*);
+void sequential_mts_double(double*,long,double*,long*);
 
 /* This function computes the maximum prefix sum in a lazy way
  * Last argument is for optional optimization method */
-void sequential_mts_lazy(double*,int,double*,int*,int);
+void sequential_mts_lazy(double*,long,double*,long*,long);
